Adjusted metadata checks to run after creating session within middleware.
