+++ app/(api)/auth/create-shop-profile/route.ts
@@ -1,7 +1,5 @@
 import { NextResponse, NextRequest } from "next/server";
 import { createProfile } from "@/utils/supabase/serverActions/createProfile";
-import { updateUser } from "@/utils/supabase/serverActions/updateUser";
-import { createClient } from "@/utils/supabase/server";
 
 export async function POST(request: NextRequest) {
 
@@ -12,13 +10,8 @@ export async function POST(request: NextRequest) {
   const { error } = await createProfile(userId);
 
   if (!error) {
+    return NextResponse.json({ message: "Account successfully created!" }, { status: 200 })
 
-    const { error: updateError } = await updateUser(userId)
-
-    if (!updateError) {
-      return NextResponse.json({ message: "Account successfully created!" }, { status: 200 })
-    }
-    return NextResponse.json({ error: updateError }, { status: 500 })
   }
 
   return NextResponse.json({ error: error }, { status: 500 })
