+++ utils/supabase/serviceClient.ts
@@ -7,7 +7,7 @@ export function createClient() {
 
   return createServerClient<Database>(
     process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.SUPABASE_SERVICE_KEY!,
+    process.env.SUPABASE_SERVICE_ROLE_KEY!,
     {
 
       cookies: {
